chore: bump version for all ee pieces
