+++ packages/ee/pieces/microsoft-dynamics-365-business-central/src/index.ts
@@ -35,7 +35,7 @@ export const microsoftDynamics365BusinessCentral = createPiece({
   displayName: 'Microsoft Dynamics 365 Business Central',
   auth: businessCentralAuth,
   categories: [PieceCategory.PREMIUM, PieceCategory.SALES_AND_CRM],
-  minimumSupportedRelease: '0.28.0',
+  minimumSupportedRelease: '0.27.1',
   logoUrl:
     'https://cdn.activepieces.com/pieces/microsoft-dynamics-365-business-central.png',
   authors: ['kishanprmr'],
