+++ packages/ee/pieces/microsoft-dynamics-crm/package.json
@@ -1,4 +1,8 @@
 {
   "name": "@activepieces/piece-microsoft-dynamics-crm",
-  "version": "0.0.4"
+  "version": "0.0.5",
+  "license": {
+    "type": "COMMERCIAL",
+    "url": "https://raw.githubusercontent.com/activepieces/activepieces/main/packages/ee/LICENSE"
+  }
 }
