+++ packages/ee/pieces/sharepoint/tsconfig.json
