+++ packages/ee/pieces/sharepoint/tsconfig.lib.json
