+++ packages/ee/pieces/zuora/src/index.ts
@@ -71,7 +71,7 @@ export const zuoraAuth = PieceAuth.CustomAuth({
 export const zuora = createPiece({
   displayName: 'Zuora',
   auth: zuoraAuth,
-  minimumSupportedRelease: '0.28.0',
+  minimumSupportedRelease: '0.27.1',
   logoUrl: 'https://cdn.activepieces.com/pieces/zuora.png',
   categories: [
     PieceCategory.SALES_AND_CRM,
