--- packages/pieces/community/microsoft-sharepoint/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@activepieces/piece-microsoft-sharepoint",
-  "version": "0.0.1"
-}
