+++ tsconfig.base.json
@@ -336,7 +336,7 @@
         "packages/pieces/community/microsoft-onedrive/src/index.ts"
       ],
       "@activepieces/piece-microsoft-sharepoint": [
-        "packages/pieces/community/microsoft-sharepoint/src/index.ts"
+        "packages/ee/pieces/sharepoint/src/index.ts"
       ],
       "@activepieces/piece-microsoft-teams": [
         "packages/pieces/community/microsoft-teams/src/index.ts"
